Moved extension to CairoExtensions
